split out building author to a util; tidied up graphql gen
